fixed lib folder, moved logger settings in config.py
